+++ app/(tabs)/settings.tsx
@@ -2,6 +2,7 @@ import ParallaxScrollView from "@/components/parallax-scroll-view";
 import ChangeThemeButton from "@/components/settings/changeThemeButton";
 import LogoutButton from "@/components/settings/logoutButton";
 import TaskFontSizeButton from "@/components/settings/taskFontSizeButton";
+import VersionInfoButton from "@/components/settings/versionInfoButton";
 import { ThemedText } from "@/components/themed-text";
 import { ThemedView } from "@/components/themed-view";
 import { IconSymbol } from "@/components/ui/icon-symbol";
@@ -44,6 +45,7 @@ export default function SettingsPage() {
 
         <ChangeThemeButton />
         <TaskFontSizeButton />
+        <VersionInfoButton />
         <LogoutButton />
       </ThemedView>
       <ThemedText style={dynamicStyles.text}>